cic(#1062): give the bottom-right zone back to the scroll pair
#997 mounted the far-behind dismiss as a third control in the #280 float
stack, FIRST in the column so it stacked above the pair. vjt navigates
far-behind windows from the mobile PWA by repeated tapping of that ONE
zone — scroll-to-bottom while there is tail below, then the mobile
next-active the instant it unmounts — so the round trip is one thumb in
one spot, and a third glyph with a third meaning lands in the middle of
it, on exactly the windows the gesture is for.

The reach argument #997 was opened on is answered elsewhere now. #1019
routes leaving a far-behind window to `dismissFarBehind`: the operator's
own window switch already unfreezes the cursor, with no control to find.
What is left of #997 — a dismiss beside the count it clears — is the
bar's ×, which is where it started and the only surface that can say
"you are N behind" in words.

The verb and its label stay: the bar is still their caller, and the
re-latch inside `dismissFarBehindGesture` is still the half an inline
handler would forget.

Tests: the #997 vitest describe asserted the old product wholesale and is
replaced by a #1062 guard that compares the stack's child count against
an ORDINARY pane — so a legitimately new permanent control does not read
as this regression, while anything the far-behind arm mounts does. The
#997 e2e spec measured the corner control's tap floor and quadrant; it is
replaced rather than deleted, because it was the only e2e that clicked a
dismiss and watched the badge fall, and that outcome moves onto the
surviving door.
